# backend/app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for external devices to access the internal Smartify system",
        "title": "Smartify Backend API",
        "contact": {
            "name": "Smartify Working Mail",
            "email": "smartifyprojectapp@gmail.com"
        },
        "version": "1.1"
    },
    "host": "213.226.112.206:22025",
    "basePath": "/api",
    "paths": {
...
        "/logout": {
            "post": {
                "description": "Deactivates the refresh token, terminating the user session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "description": "Refresh token for deactivation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Refresh_token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful exit",
                        "schema": {
                            "$ref": "#/definitions/api.Success_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
...
        "/refresh_token": {
            "post": {
                "description": "Returns a new access/refresh token pair using a valid refresh token. The old refresh token becomes invalid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "JWT Token Update",
                "parameters": [
                    {
                        "description": "Refresh token to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Refresh_token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A new pair of tokens",
                        "schema": {
                            "$ref": "#/definitions/api.Tokens_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired refresh token",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "500": {
                        "description": "Server error (token generation, database)",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
        "/registration_codevalidation": {
            "post": {
                "description": "Validates the code sent to the user's email address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registration"
                ],
                "summary": "Verifying the confirmation code",
                "parameters": [
                    {
                        "description": "Email and confirmation code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Code_verification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Code confirmed",
                        "schema": {
                            "$ref": "#/definitions/api.Success_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid code or user not found",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
        "/registration_emailvalidation": {
            "post": {
                "description": "Checks the validity of the email and sends a confirmation code. Email must not be already registered.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registration"
                ],
                "summary": "Email validation during registration",
                "parameters": [
                    {
                        "description": "User Email",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Email_struct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmation code sent",
                        "schema": {
                            "$ref": "#/definitions/api.Success_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid email or request",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "409": {
                        "description": "The user already exists",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "500": {
                        "description": "Server error (code generation, sending email)",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
        "/registration_password": {
            "post": {
                "description": "Saves the user password and issues access tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registration"
                ],
                "summary": "Finalizing registration",
                "parameters": [
                    {
                        "description": "User Email and Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.User_email_password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Access tokens",
                        "schema": {
                            "$ref": "#/definitions/api.Tokens_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid data or user not found",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "500": {
                        "description": "Server error (database, token generation)",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Code_verification": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "api.Email_struct": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "api.Error_answer": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "error": {
                    "description": "Error message",
                    "type": "string"
                }
            }
        },
...
        "api.Refresh_token": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "api.Success_answer": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "status": {
                    "description": "Status message",
                    "type": "string"
                }
            }
        },
        "api.Tokens_answer": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "api.User_email_password": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    }
}